+++ exonum-java-binding-common/src/main/java/com/exonum/binding/common/proofs/list/CheckedListProof.java
@@ -39,7 +39,7 @@ import java.util.NavigableMap;
  */
 public interface CheckedListProof<E> extends CheckedProof {
   /**
-   * Get all leaf entries of this proof.
+   * Get all list elements. There might be several consecutive ranges.
    * @throws IllegalStateException if the proof is not valid
    */
   NavigableMap<Long, E> getElements();
@@ -47,5 +47,6 @@ public interface CheckedListProof<E> extends CheckedProof {
   /**
    * Returns the status of this proof: whether it is structurally valid.
    */
+  @Override
   ListProofStatus getProofStatus();
 }