+++ exonum-java-binding-common/src/main/java/com/exonum/binding/common/proofs/list/ListProofStructureValidator.java
@@ -183,14 +183,14 @@ final class ListProofStructureValidator implements ListProofVisitor {
   /**
    * Returns proof status.
    */
-  public ListProofStatus getProofStatus() {
+  ListProofStatus getProofStatus() {
     return proofStatus;
   }
 
   /**
    * Returns true if proof status is VALID, false otherwise.
    */
-  public boolean isValid() {
+  boolean isValid() {
     return proofStatus == ListProofStatus.VALID;
   }