+++ exonum-java-binding-common/src/main/java/com/exonum/binding/common/proofs/list/UncheckedListProof.java
@@ -17,7 +17,8 @@
 package com.exonum.binding.common.proofs.list;
 
 /**
- * Interface is used to get a checked list proof.
+ * A proof that some elements exist in a proof list. You must
+ * {@link #check} its structure and root hash before accessing the elements.
  */
 public interface UncheckedListProof {