+++ exonum-java-binding-common/src/main/java/com/exonum/binding/common/proofs/map/CheckedMapProof.java
@@ -68,6 +68,7 @@ public interface CheckedMapProof extends CheckedProof {
   /**
    * Returns the status of this proof: whether it is structurally valid.
    */
+  @Override
   MapProofStatus getProofStatus();
 
   /**