+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   `fixed64`, `uint64`, `sint64`, `float` and `double` primitive types, 
   `PrivateKey`, `PublicKey` and `bytes` serialization. (#514, #523)
 - Multiproofs support in `ProofMapIndexProxy`. (#507)
+- `CheckedListProof`, `UncheckedListProof` proof interfaces and implementations. 
+  `CheckedProof` common interface. (#516)
 
 ### Changed
 - `Transaction#execute` can throw `TransactionExecutionException` to roll back 
@@ -43,6 +45,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Move `proofs` package to `com.exonum.binding.common` package. (#469)
   - Move `serialization` package to `com.exonum.binding.common` package. (#469)
 - Replace tree proof with flat proof in `ProofMapIndexProxy`. (#478)
+- `ListProofValidator` separated to `ListProofStructureValidator` 
+  and `ListProofRootHashCodeCalculator`. (#506)
 
 ### Removed
 - `Hashing#toHexString`. (#379)